+++ tcp_bsd/tcp_usrreq.c
@@ -44,7 +44,7 @@
 #include <sys/stat.h>
 
 #include <net/if.h>
-// #include <net/route.h>
+#include <net/route.h>
 
 #include <netinet/in.h>
 #include <netinet/in_systm.h>
@@ -55,8 +55,8 @@
 #include <netinet/tcp_fsm.h>
 #include <netinet/tcp_seq.h>
 #include <netinet/tcp_timer.h>
-#include <netinet/tcp_var.h>
 #include <netinet/tcpip.h>
+#include <netinet/tcp_var.h>
 #include <netinet/tcp_debug.h>
 
 /*
@@ -83,7 +83,7 @@ tcp_usrreq(so, req, m, nam, control)
 	int ostate;
 
 	if (req == PRU_CONTROL)
-		return (in_control(so, (int)m, (caddr_t)nam,
+		return (in_control(so, (long)m, (caddr_t)nam,
 			(struct ifnet *)control));
 	if (control && control->m_len) {
 		m_freem(control);
@@ -282,7 +282,7 @@ tcp_usrreq(so, req, m, nam, control)
 		}
 		m->m_len = 1;
 		*mtod(m, caddr_t) = tp->t_iobc;
-		if (((int)nam & MSG_PEEK) == 0)
+		if (((long)nam & MSG_PEEK) == 0)
 			tp->t_oobflags ^= (TCPOOB_HAVEDATA | TCPOOB_HADDATA);
 		break;
 
@@ -320,8 +320,8 @@ tcp_usrreq(so, req, m, nam, control)
 	 * routine for tracing's sake.
 	 */
 	case PRU_SLOWTIMO:
-		tp = tcp_timers(tp, (int)nam);
-		req |= (int)nam << 8;		/* for debug's sake */
+		tp = tcp_timers(tp, (long)nam);
+		req |= (long)nam << 8;		/* for debug's sake */
 		break;
 
 	default:
